tensor: support create a tensor from buffer
The background of this commit is sometimes we have a complete buffer
for tensorflow's encoded way, then we don't need to convert it to
array and then encode the array to buffer.

With this way, just pass the buffer, shape and type, there is no any
extra moving or copying tasks to run.